+++ homeassistant/components/fjaraskupan/__init__.py
@@ -1,28 +1,23 @@
 """The Fjäråskupan integration."""
 from __future__ import annotations
 
-from collections.abc import AsyncIterator, Callable
-from contextlib import asynccontextmanager
+from collections.abc import Callable
 from dataclasses import dataclass
-from datetime import timedelta
 import logging
 
-from fjaraskupan import Device, State
+from fjaraskupan import Device
 
 from homeassistant.components.bluetooth import (
     BluetoothCallbackMatcher,
     BluetoothChange,
     BluetoothScanningMode,
     BluetoothServiceInfoBleak,
-    async_address_present,
-    async_ble_device_from_address,
     async_rediscover_address,
     async_register_callback,
 )
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import Platform
 from homeassistant.core import HomeAssistant, callback
-from homeassistant.exceptions import HomeAssistantError
 from homeassistant.helpers import device_registry as dr
 from homeassistant.helpers.dispatcher import (
     async_dispatcher_connect,
@@ -30,14 +25,9 @@ from homeassistant.helpers.dispatcher import (
 )
 from homeassistant.helpers.entity import DeviceInfo, Entity
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
-from homeassistant.helpers.update_coordinator import DataUpdateCoordinator, UpdateFailed
 
 from .const import DISPATCH_DETECTION, DOMAIN
-
-
-class UnableToConnect(HomeAssistantError):
-    """Exception to indicate that we cannot connect to device."""
-
+from .coordinator import FjaraskupanCoordinator
 
 PLATFORMS = [
     Platform.BINARY_SENSOR,
@@ -50,81 +40,11 @@ PLATFORMS = [
 _LOGGER = logging.getLogger(__name__)
 
 
-class Coordinator(DataUpdateCoordinator[State]):
-    """Update coordinator for each device."""
-
-    def __init__(
-        self, hass: HomeAssistant, device: Device, device_info: DeviceInfo
-    ) -> None:
-        """Initialize the coordinator."""
-        self.device = device
-        self.device_info = device_info
-        self._refresh_was_scheduled = False
-
-        super().__init__(
-            hass, _LOGGER, name="Fjäråskupan", update_interval=timedelta(seconds=120)
-        )
-
-    async def _async_refresh(
-        self,
-        log_failures: bool = True,
-        raise_on_auth_failed: bool = False,
-        scheduled: bool = False,
-        raise_on_entry_error: bool = False,
-    ) -> None:
-        self._refresh_was_scheduled = scheduled
-        await super()._async_refresh(
-            log_failures=log_failures,
-            raise_on_auth_failed=raise_on_auth_failed,
-            scheduled=scheduled,
-            raise_on_entry_error=raise_on_entry_error,
-        )
-
-    async def _async_update_data(self) -> State:
-        """Handle an explicit update request."""
-        if self._refresh_was_scheduled:
-            if async_address_present(self.hass, self.device.address, False):
-                return self.device.state
-            raise UpdateFailed(
-                "No data received within schedule, and device is no longer present"
-            )
-
-        if (
-            ble_device := async_ble_device_from_address(
-                self.hass, self.device.address, True
-            )
-        ) is None:
-            raise UpdateFailed("No connectable path to device")
-        async with self.device.connect(ble_device) as device:
-            await device.update()
-        return self.device.state
-
-    def detection_callback(self, service_info: BluetoothServiceInfoBleak) -> None:
-        """Handle a new announcement of data."""
-        self.device.detection_callback(service_info.device, service_info.advertisement)
-        self.async_set_updated_data(self.device.state)
-
-    @asynccontextmanager
-    async def async_connect_and_update(self) -> AsyncIterator[Device]:
-        """Provide an up to date device for use during connections."""
-        if (
-            ble_device := async_ble_device_from_address(
-                self.hass, self.device.address, True
-            )
-        ) is None:
-            raise UnableToConnect("No connectable path to device")
-
-        async with self.device.connect(ble_device) as device:
-            yield device
-
-        self.async_set_updated_data(self.device.state)
-
-
 @dataclass
 class EntryState:
     """Store state of config entry."""
 
-    coordinators: dict[str, Coordinator]
+    coordinators: dict[str, FjaraskupanCoordinator]
 
 
 async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
@@ -153,7 +73,9 @@ async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
                 name="Fjäråskupan",
             )
 
-            coordinator: Coordinator = Coordinator(hass, device, device_info)
+            coordinator: FjaraskupanCoordinator = FjaraskupanCoordinator(
+                hass, device, device_info
+            )
             coordinator.detection_callback(service_info)
 
             state.coordinators[service_info.address] = coordinator
@@ -183,7 +105,7 @@ def async_setup_entry_platform(
     hass: HomeAssistant,
     entry: ConfigEntry,
     async_add_entities: AddEntitiesCallback,
-    constructor: Callable[[Coordinator], list[Entity]],
+    constructor: Callable[[FjaraskupanCoordinator], list[Entity]],
 ) -> None:
     """Set up a platform with added entities."""
 
@@ -195,7 +117,7 @@ def async_setup_entry_platform(
     )
 
     @callback
-    def _detection(coordinator: Coordinator) -> None:
+    def _detection(coordinator: FjaraskupanCoordinator) -> None:
         async_add_entities(constructor(coordinator))
 
     entry.async_on_unload(
