+++ homeassistant/components/fjaraskupan/binary_sensor.py
@@ -17,7 +17,8 @@ from homeassistant.helpers.entity import DeviceInfo, Entity
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.helpers.update_coordinator import CoordinatorEntity
 
-from . import Coordinator, async_setup_entry_platform
+from . import async_setup_entry_platform
+from .coordinator import FjaraskupanCoordinator
 
 
 @dataclass
@@ -50,7 +51,7 @@ async def async_setup_entry(
 ) -> None:
     """Set up sensors dynamically through discovery."""
 
-    def _constructor(coordinator: Coordinator) -> list[Entity]:
+    def _constructor(coordinator: FjaraskupanCoordinator) -> list[Entity]:
         return [
             BinarySensor(
                 coordinator,
@@ -64,7 +65,7 @@ async def async_setup_entry(
     async_setup_entry_platform(hass, config_entry, async_add_entities, _constructor)
 
 
-class BinarySensor(CoordinatorEntity[Coordinator], BinarySensorEntity):
+class BinarySensor(CoordinatorEntity[FjaraskupanCoordinator], BinarySensorEntity):
     """Grease filter sensor."""
 
     entity_description: EntityDescription
@@ -72,7 +73,7 @@ class BinarySensor(CoordinatorEntity[Coordinator], BinarySensorEntity):
 
     def __init__(
         self,
-        coordinator: Coordinator,
+        coordinator: FjaraskupanCoordinator,
         device: Device,
         device_info: DeviceInfo,
         entity_description: EntityDescription,
